update code to move MQTT class to grottdata and avoid sharing single object per inverter thread
